whitelinez-backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-jose[cryptography]==3.3.0
httpx==0.27.2
supabase==2.7.4
slowapi==0.1.9
structlog==24.4.0
ultralytics==8.3.40
supervision==0.25.0
opencv-python-headless==4.10.0.84
numpy==1.26.4
scipy==1.13.1
pydantic==2.10.3
python-dotenv==1.0.1
